pyproject.toml
[tool.poetry]
name = "xda"
version = "0.1.0"
description = ""
authors = ["Boy Bangkok <boybangkok@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9.1"
django = "^3.1.7"
psycopg2-binary = "^2.8.5"
whitenoise = {extras = ["brotli"], version = "^5.1.0"}
gunicorn = "^20.0.4"
django-import-export = "^2.7.1"
django-datatables-view = "^1.20.0"
django-thaidate = "^1.0.2"
djangorestframework = "^3.14.0"
django-ninja = "^0.16.1"